feat: add new playlist function (#27)
* feat: add id return on insert playlist

* refactor: make Playlist id and thumbnail url nullable

* refactor: change playlist id type to Long

* feat: implement button that adds new list

* fix: wrong checkbox behavior

* feat: add viewContainer interface

* feat: implement thumbnail auto comp

+++ app/src/main/kotlin/net/turtton/ytalarm/structure/Alarm.kt
@@ -10,7 +10,7 @@ data class Alarm(
     val id: Int? = null,
     var time: String = "00:00",
     var repeatType: RepeatType = RepeatType.ONCE,
-    var playListId: Int = 0,
+    var playListId: Long = 0,
     var loop: Boolean = false,
     // max 100(%)
     var volume: Int = 50,
